introduction.md
I am a new member of github please do help me when I'm in need of things. 
I have step my foot in the field of Analytic's with the learning knowledge in R and Python. 
Currently working in a company called KUDWI as a Business Development Analyst which is a Business Analytic company currently working on various projects in the field of Retail, Healthcare and Oil&Gas.